feat: centre of mass
Scatterplot now displays centre of mass i.e. the centre of each party's Voronoi polyhedral of voters, weighted by voter distribution. This represents the core of the party's base and will be used in calculations of the pressure the base/activists exert on the party platform.

The next step is to add in a force on the party towards this centre.